✨Enhanced `ArtifactCallback`
It will now keep artifacts for the saved checkpoints prior to others

+++ cflearn/misc/internal_/callback.py
@@ -1,4 +1,5 @@
 import os
+import json
 import mlflow
 import shutil
 import getpass
@@ -17,6 +18,8 @@ from ...trainer import Trainer
 from ...trainer import TrainerCallback
 from ...protocol import TrainerState
 from ...protocol import MetricsOutputs
+from ...constants import PT_PREFIX
+from ...constants import SCORES_FILE
 from ...constants import WARNING_PREFIX
 
 
@@ -118,7 +121,16 @@ class ArtifactCallback(TrainerCallback):
         os.makedirs(sub_folder, exist_ok=True)
         current_steps = sorted(map(int, os.listdir(sub_folder)))
         if len(current_steps) >= self.num_keep:
+            must_keep = set()
+            checkpoint_folder = trainer.checkpoint_folder
+            if checkpoint_folder is not None:
+                score_path = os.path.join(checkpoint_folder, SCORES_FILE)
+                with open(score_path, "r") as f:
+                    for key in json.load(f):
+                        must_keep.add(os.path.basename(key)[len(PT_PREFIX) :])
             for step in current_steps[: -self.num_keep + 1]:
+                if step in must_keep:
+                    continue
                 shutil.rmtree(os.path.join(sub_folder, str(step)))
         sub_folder = os.path.join(sub_folder, str(state.step))
         os.makedirs(sub_folder, exist_ok=True)
